requirements.txt
altair==5.3.0
annotated-types==0.7.0
appdirs==1.4.4
atomicwrites==1.4.0
attrs==23.2.0
backports.weakref==1.0.post1
bcbio-gff==0.7.0
biom-format==2.1.15
biopython==1.81
brotlipy==0.7.0
clyent==1.2.2
CMSeq==1.0.4
conda-repo-cli==1.0.75
conda-verify==3.4.2
constantly==15.1.0
cssselect==1.1.0
datashape==0.5.4
DendroPy==4.6.1
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
et-xmlfile==1.1.0
fastapi==0.111.1
fastapi-cli==0.0.4
ffmpy==0.3.2
fonttools==4.25.0
fsspec==2024.6.1
gradio==5.0.2
gradio_client==1.4.0
h11==0.14.0
hclust2==1.0.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.25.2
importlib_metadata==8.4.0
importlib_resources==6.4.0
inflection==0.5.1
jiter==0.5.0
jsonpatch==1.33
jsonpointer==2.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
langchain-core==0.2.22
langchain-openai==0.1.17
langsmith==0.1.93
litellm==1.44.22
load-dotenv==0.1.0
MetaPhlAn==4.0.6
munkres==1.1.4
openai==1.44.0
openpyxl==3.0.10
orjson==3.10.6
outcome==1.3.0.post0
packaging==24.1
pandarallel==1.6.5
patsy==0.5.3
pep8==1.7.1
PhyloPhlAn==3.0.3
ply==3.11
pyarrow==11.0.0
pyasn1-modules==0.2.8
pydantic==2.8.2
pydantic_core==2.20.1
PyDispatcher==2.0.5
pydub==0.25.1
pyls-spyder==0.4.0
PyQt5-sip==12.11.0
pysam==0.22.0
python-lsp-jsonrpc==1.0.0
python-multipart==0.0.9
quast==5.2.0
queuelib==1.5.0
referencing==0.35.1
requests==2.32.3
rich==13.7.1
routellm==0.2.0
rpds-py==0.20.0
ruff==0.5.4
selenium==4.21.0
semantic-version==2.10.0
sglang==0.3.0
shellingham==1.5.4
shortuuid==1.0.13
simplejson==3.19.2
sniffio==1.3.1
SpeechRecognition==3.10.1
starlette==0.37.2
tiktoken==0.7.0
tomlkit==0.12.0
torch==2.4.1
trio==0.25.1
trio-websocket==0.11.1
typer==0.12.3
typing_extensions==4.11.0
urllib3==2.2.2
uvicorn==0.30.3
uvloop==0.19.0
watchfiles==0.22.0
webdriver-manager==4.0.1
webencodings==0.5.1
websockets==11.0.3
wsproto==1.2.0
